fix: update comments
