+++ Architect_MultiClient_Server/orchestrator_service/.env.example
@@ -129,3 +129,9 @@ NOTIFICATION_MAX_RETRIES="3"
 NOTIFICATION_RETRY_DELAY_SEC="5"
 NOTIFICATION_CHANNEL_ID=""
 NOTIFICATION_WEBHOOK_TOKEN=""
+
+# =========================================================
+# OUTBOX_CONFIG
+# =========================================================
+OUTBOX_RETRY_SUMMARIZATION_TARGET_HOURS=19,20,21
+OUTBOX_RETRY_SUMMARIZATION_LIMIT=5
