+++ Architect_MultiClient_Server/orchestrator_service/migrations_pg/versions/004_add_outbox_tasks_with_enums.py
@@ -1,8 +1,8 @@
-"""add outbox tasks table
+"""add outbox tasks table with enums
 
-Revision ID: 004_add_outbox_tasks
+Revision ID: 004_add_outbox_tasks_with_enums
 Revises: 003_drop_full_text_column
-Create Date: 2026-06-01 12:00:00.000000
+Create Date: 2026-06-02 15:00:00.000000
 
 """
 
@@ -10,21 +10,25 @@ from alembic import op
 import sqlalchemy as sa
 from sqlalchemy.dialects import postgresql
 
-
 # revision identifiers, used by Alembic.
-revision = '004_add_outbox_tasks'
+revision = '004_add_outbox_tasks_with_enums'
 down_revision = '003_drop_full_text_column'
 branch_labels = None
 depends_on = None
 
 
 def upgrade() -> None:
-    # Create outbox_tasks table
+    use_case_enum = postgresql.ENUM('retry_summarization', name='outboxusecase')
+    use_case_enum.create(op.get_bind(), checkfirst=True)
+    
+    status_enum = postgresql.ENUM('pending', 'processing', 'completed', 'failed', name='outboxstatus')
+    status_enum.create(op.get_bind(), checkfirst=True)
+
     op.create_table(
         'outbox_tasks',
         sa.Column('id', sa.UUID(), nullable=False),
-        sa.Column('use_case', sa.Text(), nullable=False),
-        sa.Column('status', sa.Text(), nullable=False, server_default='pending'),
+        sa.Column('use_case', use_case_enum, nullable=False),
+        sa.Column('status', status_enum, nullable=False, server_default='pending'),
         sa.Column('configs', postgresql.JSONB(), nullable=False, server_default='{}'),
         sa.Column('last_error', sa.Text(), nullable=True),
         sa.Column('created_at', sa.DateTime(timezone=True), nullable=False, server_default=sa.func.now()),
@@ -32,7 +36,6 @@ def upgrade() -> None:
         sa.PrimaryKeyConstraint('id')
     )
 
-    # Create Indexes
     op.create_index(
         'ix_outbox_tasks_status_created_at',
         'outbox_tasks',
@@ -46,9 +49,9 @@ def upgrade() -> None:
 
 
 def downgrade() -> None:
-    # Drop Indexes
     op.drop_index('ix_outbox_tasks_status_created_at', table_name='outbox_tasks')
     op.drop_index('ix_outbox_tasks_use_case', table_name='outbox_tasks')
 
-    # Drop table
     op.drop_table('outbox_tasks')
+    op.execute("DROP TYPE outboxusecase")
+    op.execute("DROP TYPE outboxstatus")
