+++ Architect_MultiClient_Server/orchestrator_service/services/postgresql/pg_transcript_repository.py
@@ -12,8 +12,6 @@ from sqlalchemy import text
 from orchestrator_service.services.postgresql.database import get_session_factory
 from orchestrator_service.utils.logger import get_logger
 from orchestrator_service.utils.time_convert import convert_to_iso_8601
-from orchestrator_service.services.postgresql.models import OutboxUseCase, OutboxStatus
-from orchestrator_service.models.summary_models import RetryType
 
 logger = get_logger(__name__)
 
@@ -1298,149 +1296,3 @@ class PgTranscriptRepository:
         except Exception as e:
             logger.error(f"Failed to list tracks: {e}")
             return []
-
-    # ------------------------------------------------------------------
-    # GENERIC OUTBOX TASKS
-    # ------------------------------------------------------------------
-
-    async def add_retry_summarization_task_to_outbox(
-        self,
-        room_id: str,
-        retry_type: str,
-        error_msg: str,
-    ) -> bool:
-        session_factory = get_session_factory()
-        now = datetime.now(timezone.utc)
-        configs = {"room_id": room_id, "retry_type": retry_type}
-        configs_json = json.dumps(configs)
-
-        try:
-            async with session_factory() as session:
-                # Check for existing pending task for same room_id and retry_type
-                check_query = """
-                    SELECT id FROM outbox_tasks 
-                    WHERE use_case = :use_case
-                      AND configs->>'room_id' = :room_id 
-                      AND retry_type = :retry_type
-                      AND status = :status
-                    LIMIT 1
-                """
-                res = await session.execute(
-                    text(check_query),
-                    {
-                        "use_case": OutboxUseCase.RETRY_SUMMARIZATION.value,
-                        "room_id": room_id, 
-                        "retry_type": retry_type,
-                        "status": OutboxStatus.PENDING.value
-                    }
-                )
-                existing = res.fetchone()
-
-                if existing:
-                    # Update existing pending task's error and timestamp
-                    task_id = existing[0]
-                    update_query = """
-                        UPDATE outbox_tasks 
-                        SET last_error = :error_msg,
-                            updated_at = :now
-                        WHERE id = :id
-                    """
-                    await session.execute(
-                        text(update_query),
-                        {
-                            "id": task_id,
-                            "error_msg": error_msg,
-                            "now": now,
-                        }
-                    )
-                else:
-                    # Insert new outbox task
-                    task_id = str(uuid.uuid4())
-                    insert_query = """
-                        INSERT INTO outbox_tasks 
-                            (id, use_case, status, retry_type, configs, last_error, created_at, updated_at)
-                        VALUES
-                            (:id, :use_case, :status, :retry_type, :configs::jsonb, :error_msg, :now, :now)
-                    """
-                    await session.execute(
-                        text(insert_query),
-                        {
-                            "id": task_id,
-                            "use_case": OutboxUseCase.RETRY_SUMMARIZATION.value,
-                            "status": OutboxStatus.PENDING.value,
-                            "retry_type": retry_type,
-                            "configs": configs_json,
-                            "error_msg": error_msg,
-                            "now": now,
-                        }
-                    )
-                await session.commit()
-                return True
-        except Exception as e:
-            logger.error(f"Failed to add task to outbox: {e}")
-            return False
-
-    async def fetch_pending_outbox_tasks(self, limit: int = 5, use_case: Optional[str] = None) -> List[Dict[str, Any]]:
-        from orchestrator_service.services.postgresql.models import OutboxStatus, OutboxUseCase
-        session_factory = get_session_factory()
-
-        query = """
-            SELECT * FROM outbox_tasks 
-            WHERE status = :status
-        """
-        params = {"limit": limit, "status": OutboxStatus.PENDING.value}
-
-        if use_case:
-            query += " AND use_case = :use_case"
-            params["use_case"] = use_case
-
-        query += """
-            ORDER BY created_at ASC 
-            LIMIT :limit 
-            FOR UPDATE SKIP LOCKED
-        """
-
-        try:
-            async with session_factory() as session:
-                res = await session.execute(
-                    text(query),
-                    params
-                )
-                rows = res.fetchall()
-                tasks = []
-                for row in rows:
-                    t = dict(row._mapping)
-                    t["id"] = str(t["id"])
-                    tasks.append(t)
-                return tasks
-        except Exception as e:
-            logger.error(f"Failed to fetch pending outbox tasks: {e}")
-            return []
-
-    async def update_outbox_task_status(
-        self,
-        task_id: str,
-        status: str,
-        error_msg: Optional[str] = None,
-    ) -> bool:
-        session_factory = get_session_factory()
-        now = datetime.now(timezone.utc)
-
-        # Status is passed as a string/enum value from worker
-        query = "UPDATE outbox_tasks SET status = :status, updated_at = :now"
-        params = {"id": task_id, "status": status, "now": now}
-
-        if error_msg is not None:
-            query += ", last_error = :error_msg"
-            params["error_msg"] = error_msg
-
-        query += " WHERE id = :id"
-
-        try:
-            async with session_factory() as session:
-                await session.execute(text(query), params)
-                await session.commit()
-                return True
-        except Exception as e:
-            logger.error(f"Failed to update outbox task {task_id}: {e}")
-            return False
